requirements.txt
fastapi
uvicorn
google-generativeai
markdown
weasyprint
python-dotenv
pytest
flake8